+++ webapp/app/assessment/components/description/actions.js
