+++ webapp/app/assessment/components/description/commentableDescription.js
