+++ webapp/app/assessment/components/description/dataSourceDescriptionAndComments.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import { useSelector } from 'react-redux'
 import * as R from 'ramda'
 import ReviewIndicator from '@webapp/app/assessment/components/review/reviewIndicator'
-import Description from '@webapp/components/description/description'
+import Description from '@webapp/app/assessment/components/description/description'
 import * as ReviewState from '@webapp/app/assessment/components/review/reviewState'
 import useCountryIso from '@webapp/components/hooks/useCountryIso'
 import useI18n from '@webapp/components/hooks/useI18n'
