+++ webapp/app/assessment/components/description/description.less
