+++ webapp/app/assessment/components/description/descriptionState.js
