+++ webapp/app/assessment/components/description/nationalDataDescriptions.js
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { isPrintingMode, isPrintingOnlyTables } from '@webapp/app/assessment/components/print/printAssessment'
 
-import CommentableDescription from '@webapp/components/description/commentableDescription'
+import CommentableDescription from '@webapp/app/assessment/components/description/commentableDescription'
 import useI18n from '@webapp/components/hooks/useI18n'
 
 const NationalDataDescriptions = props => {
