+++ webapp/app/assessment/fra/sections/biomassStock/biomassStockView.js
@@ -6,9 +6,9 @@ import TraditionalTable from '@webapp/app/assessment/components/traditionalTable
 import tableSpec from './tableSpec'
 import DefinitionLink from '@webapp/components/definitionLink'
 import { fetchLastSectionUpdateTimestamp } from '@webapp/app/components/audit/actions'
-import NationalDataDescriptions from '@webapp/components/description/nationalDataDescriptions'
-import AnalysisDescriptions from '@webapp/components/description/analysisDescriptions'
-import GeneralComments from '@webapp/components/description/generalComments'
+import NationalDataDescriptions from '@webapp/app/assessment/components/description/nationalDataDescriptions'
+import AnalysisDescriptions from '@webapp/app/assessment/components/description/analysisDescriptions'
+import GeneralComments from '@webapp/app/assessment/components/description/generalComments'
 import ExcelCalculatorDownload from '@webapp/app/assessment/fra/sections/biomassStock/excelCalculatorDownload'
 import { isFRA2020SectionEditDisabled } from '@webapp/utils/assessmentAccess'
 import * as table from '@webapp/app/assessment/components/traditionalTable/table'
