+++ webapp/app/assessment/fra/sections/extentOfForest/extentOfForestView.js
@@ -6,9 +6,9 @@ import Icon from '@webapp/components/icon'
 import DefinitionLink from '@webapp/components/definitionLink'
 import ChartWrapper from './chart/chartWrapper'
 import { GenerateFraValuesControl, TableWithOdp } from '../../components/tableWithOdp/tableWithOdp'
-import NationalDataDescriptions from '@webapp/components/description/nationalDataDescriptions'
-import AnalysisDescriptions from '@webapp/components/description/analysisDescriptions'
-import GeneralComments from '@webapp/components/description/generalComments'
+import NationalDataDescriptions from '@webapp/app/assessment/components/description/nationalDataDescriptions'
+import AnalysisDescriptions from '@webapp/app/assessment/components/description/analysisDescriptions'
+import GeneralComments from '@webapp/app/assessment/components/description/generalComments'
 import ReviewIndicator from '@webapp/app/assessment/components/review/reviewIndicator'
 import TraditionalTable from '@webapp/app/assessment/components/traditionalTable/traditionalTable'
 import NationalDataPointsPrintView from '../originalDataPoint/nationalDataPointsPrintView'
