+++ webapp/app/assessment/fra/sections/forestCharacteristics/forestCharacteristicsView.js
@@ -10,9 +10,9 @@ import { getForestAreaForYear, hasOdps } from '@common/extentOfForestHelper'
 
 import { GenerateFraValuesControl,TableWithOdp } from '@webapp/app/assessment/fra/components/tableWithOdp/tableWithOdp'
 import ChartWrapper from '@webapp/app/assessment/fra/sections/extentOfForest/chart/chartWrapper'
-import NationalDataDescriptions from '@webapp/components/description/nationalDataDescriptions'
-import AnalysisDescriptions from '@webapp/components/description/analysisDescriptions'
-import GeneralComments from '@webapp/components/description/generalComments'
+import NationalDataDescriptions from '@webapp/app/assessment/components/description/nationalDataDescriptions'
+import AnalysisDescriptions from '@webapp/app/assessment/components/description/analysisDescriptions'
+import GeneralComments from '@webapp/app/assessment/components/description/generalComments'
 import DefinitionLink from '@webapp/components/definitionLink'
 import ReviewIndicator from '@webapp/app/assessment/components/review/reviewIndicator'
 import NationalDataPointsPrintView from '@webapp/app/assessment/fra/sections/originalDataPoint/nationalDataPointsPrintView'
