+++ webapp/app/assessment/fra/sections/growingStock/growingStockView.js
@@ -13,9 +13,9 @@ import ButtonTableExport from '@webapp/components/buttonTableExport'
 import { ThousandSeparatedDecimalInput } from '@webapp/components/thousandSeparatedDecimalInput'
 import DefinitionLink from '@webapp/components/definitionLink'
 import ReviewIndicator from '@webapp/app/assessment/components/review/reviewIndicator'
-import NationalDataDescriptions from '@webapp/components/description/nationalDataDescriptions'
-import AnalysisDescriptions from '@webapp/components/description/analysisDescriptions'
-import GeneralComments from '@webapp/components/description/generalComments'
+import NationalDataDescriptions from '@webapp/app/assessment/components/description/nationalDataDescriptions'
+import AnalysisDescriptions from '@webapp/app/assessment/components/description/analysisDescriptions'
+import GeneralComments from '@webapp/app/assessment/components/description/generalComments'
 import useCountryIso from '@webapp/components/hooks/useCountryIso'
 import useUserInfo from '@webapp/components/hooks/useUserInfo'
