+++ webapp/app/assessment/fra/sections/nonWoodForestProductsRemovals/nonWoodForestProductsRemovalsView.js
@@ -8,8 +8,8 @@ import TraditionalTable from '@webapp/app/assessment/components/traditionalTable
 import mainTableSpec, { sectionName } from './mainTableSpec'
 import DefinitionLink from '@webapp/components/definitionLink'
 import { fetchLastSectionUpdateTimestamp } from '@webapp/app/components/audit/actions'
-import NationalDataDescriptions from '@webapp/components/description/nationalDataDescriptions'
-import GeneralComments from '@webapp/components/description/generalComments'
+import NationalDataDescriptions from '@webapp/app/assessment/components/description/nationalDataDescriptions'
+import GeneralComments from '@webapp/app/assessment/components/description/generalComments'
 import { isFRA2020SectionEditDisabled } from '@webapp/utils/assessmentAccess'
 import { fetchTableData } from '@webapp/app/assessment/components/traditionalTable/actions'
 import * as table from '@webapp/app/assessment/components/traditionalTable/table'
